Broke out the promotion definition
